Abstracted UpdateTest to TestCaseUtils
