+++ master/src/test/java/com/examples/with/different/packagename/dse/ArrayLengthExample.java
@@ -5,16 +5,24 @@ public class ArrayLengthExample {
 	public ArrayLengthExample() {
 	}
 
+//	public static int max(int[] array) {
+//		new ArrayLengthExample();
+//		if (array==null) {
+//			return -1;
+//		} else if (array.length==0) {
+//			return 0;
+//		} else if (array.length==1) {
+//			return 1;
+//		} else {
+//			return array.length;
+//		}
+//	}
+
 	public static int max(int[] array) {
-		new ArrayLengthExample();
-		if (array==null) {
+		if (array.length == 0) {
 			return -1;
-		} else if (array.length==0) {
-			return 0;
-		} else if (array.length==1) {
-			return 1;
 		} else {
-			return array.length;
+			return 0;
 		}
 	}
